refactor: move baseentity to common
