+++ backend/src/main/java/team/teamby/teambyteam/icalendar/domain/PublishedIcalendar.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import team.teamby.teambyteam.global.domain.BaseEntity;
+import team.teamby.teambyteam.common.domain.BaseEntity;
 import team.teamby.teambyteam.icalendar.domain.vo.IcalendarFileName;
 import team.teamby.teambyteam.icalendar.domain.vo.PublishUrl;
