+++ backend/src/main/java/team/teamby/teambyteam/teamplace/domain/TeamPlace.java
@@ -8,7 +8,7 @@ import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import team.teamby.teambyteam.global.domain.BaseEntity;
+import team.teamby.teambyteam.common.domain.BaseEntity;
 import team.teamby.teambyteam.teamplace.domain.vo.Name;
 
 @Getter
